fix(challenges): Small improvement to d3 scale lesson
When I was studying https://learn.freecodecamp.org/data-visualization/data-visualization-with-d3/use-a-pre-defined-scale-to-place-elements I noticed that the challenge forces the student to solve it in an undesirable way. If you desire the text label to be offset 10px from the center of the data point that it labels, then you should add 10px after the scaling function, not before. Otherwise, using the data in this challenge, you will have some labels that are offset about 10px to the right of the data point and some that are closer to 9px. This problem would have been more noticeable if it wasn’t for the coincidence that the scaling function has a scaling multiplier that is really close to x1.